Add setup & tf files for Cloud Functions
