+++ tf/cloud-function/cloud-function.tf
@@ -1,9 +1,23 @@
+# Copyright 2021 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 resource "google_cloudfunctions2_function" "function" {
   name     = "e2etest-${terraform.workspace}"
   location = "us-central1"
 
   timeouts {
-    create = "1m"
+    create = "5m"
   }
 
   build_config {
@@ -21,27 +35,34 @@ resource "google_cloudfunctions2_function" "function" {
     max_instance_count = 3
     min_instance_count = 1
     available_memory = "256M"
-    timeout_seconds = 600
+    timeout_seconds = 540
     environment_variables = {
       # Environment variables available during function execution
       "PROJECT_ID" = var.project_id
-      "REQUEST_SUBSCRIPTION_NAME" = module.pubsub.info.request_topic.subscription_name
-      "RESPONSE_TOPIC_NAME" = module.pubsub.info.response_topic.topic_name        
+      "RESPONSE_TOPIC_NAME" = module.pubsub.info.response_topic.topic_name
     }
     ingress_settings = "ALLOW_INTERNAL_ONLY"
-    all_traffic_on_latest_revision = true
+    all_traffic_on_latest_revision = true    
   }
 
   event_trigger {
     trigger_region = "us-central1"
     event_type = "google.cloud.pubsub.topic.v1.messagePublished"
-    pubsub_topic =  module.pubsub.info.request_topic.topic_name
-    retry_policy = "RETRY_POLICY_RETRY"
+    pubsub_topic =  google_pubsub_topic.topic.id
+    retry_policy = "RETRY_POLICY_RETRY"    
   }
 }
 
+resource "google_pubsub_topic" "topic" {
+  name = "cloudfunctions-push-topic"
+}
+
+resource "random_id" "bucket_prefix" {
+  byte_length = 8
+}
+
 resource "google_storage_bucket" "bucket" {
-  name                        = "opentelemetry-e2e-test-gcf-source" # Every bucket name must be globally unique
+  name                        = "${random_id.bucket_prefix.hex}-gcf-source" # Every bucket name must be globally unique
   location                    = "US"
   uniform_bucket_level_access = true
 }
@@ -49,7 +70,7 @@ resource "google_storage_bucket" "bucket" {
 resource "google_storage_bucket_object" "object" {
   name   = "function-source.zip"
   bucket = google_storage_bucket.bucket.name
-  source = "${var.runtime}/${var.source}" # Add path to the zipped function source code, source file zip should be in the bucket
+  source = "${var.functionsource}" # Add path to the zipped function source code, source file zip should be in the bucket
 }
 
 module "pubsub" {
@@ -58,11 +79,26 @@ module "pubsub" {
   project_id = var.project_id
 }
 
-variable "image" {
+variable "runtime" {
+  type = string
+}
+
+variable "entrypoint" {
+  type = string
+}
+
+variable "functionsource" {
   type = string
 }
 
 output "pubsub_info" {
-  value       = module.pubsub.info
+  value       = {
+    request_topic = {
+      topic_name = google_pubsub_topic.topic.name
+      subscription_name = "n/a" # subscription name is not required for cloud functions
+    }
+
+    response_topic = module.pubsub.info.response_topic
+  }
   description = "Info about the request/response pubsub topics and subscription to use in the test"
 }
